requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
cffi==1.14.2
click==7.1.2
dj-database-url==0.5.0
Django==3.1.1
django-braces==1.14.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
flake8==3.8.3
libsass==0.20.1
mccabe==0.6.1
misaka==2.1.1
mypy-extensions==0.4.3
pathspec==0.8.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
